Small changes to pipelines - Moved CompletionData to another file and not as a nested type of Pipe - Follow the file name guidelines for tfm specific implementations
